@@ -23,8 +23,8 @@ package org.birthdayadapter.ui;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v7.preference.Preference;
-import android.support.v7.preference.PreferenceFragmentCompat;
+import androidx.preference.Preference;
+import androidx.preference.PreferenceFragmentCompat;
 
 import org.birthdayadapter.BuildConfig;
 import org.birthdayadapter.R;
