@@ -26,8 +26,8 @@ import android.content.Context;
 import android.content.DialogInterface;
 import android.content.res.TypedArray;
 import android.os.Build;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.preference.Preference;
+import androidx.appcompat.app.AlertDialog;
+import androidx.preference.Preference;
 import android.text.format.DateFormat;
 import android.util.AttributeSet;
 import android.view.LayoutInflater;
