@@ -23,7 +23,7 @@ package org.birthdayadapter.util;
 import android.accounts.Account;
 import android.content.Context;
 import android.os.Build;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
